+++ sertit/rasters.py
@@ -1696,8 +1696,12 @@ def read_bit_array(
                [1, 0, 0]], dtype=uint8)
     """
     # TODO: daskify this, should be straightforward if unpackbits is daskified
-    if isinstance(bit_mask, xr.DataArray):
-        bit_mask = bit_mask.data
+    if isinstance(bit_mask, np.ndarray):
+        bit_mask = np.nan_to_num(bit_mask)
+    elif isinstance(bit_mask, xr.DataArray):
+        bit_mask = bit_mask.fillna(0).data
+    else:
+        bit_mask = bit_mask.fillna(0)
 
     return rasters_rio.read_bit_array(bit_mask, bit_id)
 
@@ -1720,7 +1724,14 @@ def read_uint8_array(
     Returns:
         Union[np.ndarray, list]: Binary mask or list of binary masks if a list of bit_id is given
     """
-    return read_bit_array(bit_mask.fillna(0).astype(np.uint8), bit_id)
+    if isinstance(bit_mask, np.ndarray):
+        bit_mask = np.nan_to_num(bit_mask)
+    elif isinstance(bit_mask, xr.DataArray):
+        bit_mask = bit_mask.fillna(0).data
+    else:
+        bit_mask = bit_mask.fillna(0)
+
+    return read_bit_array(bit_mask.astype(np.uint8), bit_id)
 
 
 def set_metadata(